Issues #72 and #82
The mean calculation for count now includes an offset to handle events that have partial plots (eg NETN in 2006 only had 1 microplot for seedlings and saplings). The maximum fitted.value is the mean (same as the exp(coef) if offsets are the same).

Since the size group for non-herbs is already/ha, I updated the code in line 93 to only convert to /ha for counts.

Also fixed multiple spelling errors